Move CatalogItemSelectorModal to shared components folder
